requirements.txt
astroid==2.1.0
colorama==0.4.1
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.5
django-import-export==1.2.0
et-xmlfile==1.0.1
isort==4.3.4
jdcal==1.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
odfpy==1.4.0
openpyxl==2.5.14
pylint==2.2.2
pytz==2018.7
PyYAML==3.13
six==1.12.0
tablib==0.12.1
typed-ast==1.1.1
unicodecsv==0.14.1
wrapt==1.10.11
xlrd==1.2.0
xlwt==1.3.0
